[roottest] Disable tests that require NumPy if NumPy is not available
Disable tests that require NumPy if NumPy is not available by
factorizing the relevant tests out into a separate tests, where the
NumPy dependency is correctly registered.

This helps when testing ROOT with newer Python versions where NumPy
is not readily available yet.
